Extract AttributesImpacted to its own component
